src/requirements.txt
Django=3.0.3
django-crispy-forms=1.8.1
Pillow=7.0.0